feat: stacks history api (#216)
* feat: Stacks history utils

* chore: backward compatible export

---------
