+++ types/api/stacks/index.ts
@@ -1,2 +1,3 @@
 export * from './assets';
 export * from './transaction';
+export * from './history';
